+++ CHANGELOG.md
@@ -8,7 +8,7 @@
 - Fix image sources not being marked as loaded on error
 - Fix ScaleControl options should be optional. ([#4002](https://github.com/maplibre/maplibre-gl-js/pull/4002))
 - Fix race condition in SourceCache that makes unit tests unstable. Eliminate a redundant 'visibility' event fired from Style class. ([#3992](https://github.com/maplibre/maplibre-gl-js/issues/3992))
-
+- Fix MapOptions#style is optional in the Map constructor([#4003](https://github.com/maplibre/maplibre-gl-js/pull/4003))
 
 ## 4.1.2